+++ Controls-demo/list_new/ItemClick/Base/ItemClick.wml
@@ -1,0 +1,13 @@
+<div class="controlsDemo__wrapper">
+    <div name="currentItem" style="border: 1px solid red; width: 256px;">{{_currentItem ? 'Выполнили клик по ' + _currentItem : ''}}</div>
+    <Controls.list:View
+            attr:class="controlsDemo__margin-top"
+            keyProperty="id"
+            multiSelectVisibility="{{ _hasMultiSelect ? 'visible' : 'hidden'}}"
+            source="{{_viewSource}}"
+            itemActions="{{_itemActions}}"
+            bind:selectedKeys="_selectedKeys"
+            on:itemClick="_onItemClick()"
+            on:itemActivate="_onItemActivate()"
+    />
+</div>
